+++ packages/webview_flutter/webview_flutter_android/lib/webview_android_widget.dart
@@ -3,8 +3,10 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:typed_data';
 
 import 'package:flutter/widgets.dart';
+import 'package:webview_flutter_android/webview_android_cookie_manager.dart';
 import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface.dart';
 
 import 'src/android_webview.dart' as android_webview;
@@ -210,6 +212,28 @@ class WebViewAndroidPlatformController extends WebViewPlatformController {
     return webView.loadUrl(url, headers ?? <String, String>{});
   }
 
+  /// When making a POST request, headers are ignored. As a workaround, make
+  /// the request manually and load the response data using [loadHTMLString].
+  @override
+  Future<void> loadRequest(
+    WebViewRequest request,
+  ) async {
+    if (!request.uri.hasScheme) {
+      throw ArgumentError('WebViewRequest#uri is required to have a scheme.');
+    }
+    switch (request.method) {
+      case WebViewRequestMethod.get:
+        return webView.loadUrl(request.uri.toString(), request.headers);
+      case WebViewRequestMethod.post:
+        return webView.postUrl(
+            request.uri.toString(), request.body ?? Uint8List(0));
+      default:
+        throw UnimplementedError(
+          'This version of webview_android_widget currently has no implementation for HTTP method ${request.method.serialize()} in loadRequest.',
+        );
+    }
+  }
+
   @override
   Future<String?> currentUrl() => webView.getUrl();
 
@@ -340,6 +364,14 @@ class WebViewAndroidPlatformController extends WebViewPlatformController {
     }
 
     addJavascriptChannels(creationParams.javascriptChannelNames);
+
+    // TODO(BeMacized): Remove once platform implementations
+    // are able to register themselves (Flutter >=2.8),
+    // https://github.com/flutter/flutter/issues/94224
+    WebViewCookieManagerPlatform.instance ??= WebViewAndroidCookieManager();
+
+    creationParams.cookies
+        .forEach(WebViewCookieManagerPlatform.instance!.setCookie);
   }
 
   Future<void> _setHasProgressTracking(bool hasProgressTracking) async {