+++ packages/video_player/video_player/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Updates unit tests to be self-contained.
 * Fixes integration tests.
 * Updates Android compileSdkVersion to 31.
+* Fixes a flaky integration test.
 
 ## 2.2.7