+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.7.1
+
+* Fixes header import for cookie manager to be relative only. 
+
+## 2.7.0
+
+* Adds implementation of the `loadFlutterAsset` method from the platform interface.
+
+## 2.6.0
+
+* Implements new cookie manager for setting cookies and providing initial cookies.
+
 ## 2.5.0
 
 * Adds an option to set the background color of the webview.